wip: some organisation and comments
